+++ src/shared/layout/footer/footer.jsx
@@ -1,9 +1,9 @@
 import { Link } from 'react-router-dom';
 export default function Footer () {
   return (
-    <footer className="fixed bottom-0 w-full flex justify-around items-center py-4">
+    <footer className="fixed bottom-0 w-full flex justify-around items-center py-4 text-xs">
       <p>© 2025 Agent</p>
-      <div className="flex flex-col items-center gap-4">
+      <div className="flex flex-col md:flex-row items-center gap-4">
         <Link to="/policy/privacy">
           <span className="">Privacy Policy</span>
         </Link>
